Allow onload callbacks to be threaded
